requirements.txt
feedparser>=6.0.10
beautifulsoup4>=4.12.0
pandas>=2.0.0
trafilatura>=1.6.0
lxml>=4.9.0
lxml_html_clean>=0.4.0
pymorphy3>=2.0.0
sentence-transformers>=2.2.0
sentencepiece>=0.2.0
protobuf>=3.20.0
transformers>=4.30.0,<5.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.0.0
pytest>=7.4.0
gigachat>=0.1.0
langgraph>=0.0.20
openpyxl>=3.1.0
fastapi>=0.104.0
email-validator>=2.0.0
uvicorn>=0.24.0
httpx>=0.27.0
# Примечание: используем gigachat напрямую, НЕ langchain-gigachat
# langchain-gigachat конфликтует с langgraph (требует старую версию langchain-core)

# BERTopic (карта тем)
bertopic>=0.16.0
hdbscan>=0.8.33
umap-learn>=0.5.0

# PostgreSQL
psycopg2-binary>=2.9.0

# Auth (JWT + password hashing)
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6

